refactor(daemon): split storage_decision_sweep_cmd by concern (1317→586)
The sweep command file bundled six concerns in one 1317-line unit.
Split into cohesive sibling files within package main — pure
relocation, no call-site or behavior changes:

- storage_decision_sweep_cmd.go (586): command entry (run/usage) +
  execute orchestration and its pipeline helpers
- storage_decision_sweep_types.go: DTO/value types + defaults
- storage_decision_sweep_deps.go: recordings-service / deps / truth-
  source / audit-sink adapters
- storage_decision_sweep_diff.go: state load/persist + run-to-run
  diffing and change collectors
- storage_decision_sweep_render.go: table rendering
- storage_decision_sweep_clients.go: client-profile selection + caps

45 funcs in, 45 funcs out (lossless). Verified by go build, go vet,
cmd/daemon tests, scoped golangci-lint, goimports; config surfaces
unchanged.
